requirements.txt
watchdog>=3.0.0
wmi>=1.5.1; platform_system == "Windows"
pywin32>=306; platform_system == "Windows"
pydub>=0.25.1
whisperx>=3.0.0
openai-whisper>=1.0.0
ffmpeg-python>=0.2.0
mutagen>=1.46.0
pystray>=0.19.4; platform_system == "Windows"
rumps>=0.4.0; platform_system == "Darwin"
PyGObject>=3.44.1; platform_system == "Linux"
pyudev>=0.24.1; platform_system == "Linux"
Pillow>=10.0.0